added props for button and move interfaces
