+++ frontend/src/Components/Atoms/Button/Button.styles.ts
@@ -1,9 +1,10 @@
 import styled from '@emotion/styled'
+import { StyleButtonProps } from './interface'
 
-export const StyledButton = styled.button`
-    min-width: 100px;
-    min-height: 40px;
-    font-size: 15px;
+export const StyledButton = styled.button<StyleButtonProps>`
+    width: ${(props) => props.width}px;
+    height: ${(props) => props.height}px;
+    font-size: ${(props) => props.size}px;
     border: none;
     cursor: pointer;
 `
