+++ frontend/src/Components/Atoms/Button/__tests__/Button.spec.tsx
@@ -1,5 +1,5 @@
 import { render, fireEvent, screen } from '@testing-library/react'
-import { Button } from './Button'
+import { Button } from '../Button'
 
 test('<Button />', async () => {
     render(<Button children="test" />)
