+++ frontend/src/Components/Atoms/Button/interface.ts
@@ -1,0 +1,14 @@
+import { ReactNode } from 'react'
+
+export interface StyleButtonProps {
+    height?: number
+    width?: number
+    size?: number
+}
+
+export interface IButtonProps {
+    children: ReactNode
+}
+
+export type IProps = IButtonProps & StyleButtonProps
+
