+++ frontend/src/Components/Molecules/AuthForm/interface.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export interface IFormProps {
+    children: ReactNode
+}
